feat: finish create order request

+++ src/main/kotlin/com/sample/webfluxblockingdatabase/core/domain/dto/TransactionRequestDto.kt
@@ -2,5 +2,5 @@ package com.sample.webfluxblockingdatabase.core.domain.dto
 
 data class TransactionRequestDto(
     val userId: Int,
-    val amount: Int
+    val amount: Double
 )
