+++ src/main/kotlin/com/sample/webfluxblockingdatabase/core/domain/dto/UserDto.kt
@@ -3,5 +3,5 @@ package com.sample.webfluxblockingdatabase.core.domain.dto
 data class UserDto(
     val id: Int,
     val name: String,
-    val balance: Int
+    val balance: Double
 )
